added linting, minimal testing, and github actions
