+++ .eslintrc.json
@@ -1,0 +1,18 @@
+{
+    "env": {
+        "browser": true,
+        "commonjs": true,
+        "es6": true
+    },
+    "extends": ["airbnb-base", "prettier"],
+    "plugins": ["prettier"],
+    "rules": {
+      "prettier/prettier": ["error"],
+      "no-console": [
+        "warn",
+        {
+          "allow": ["warn", "info", "error"]
+        }
+      ]
+    }
+}
