+++ .github/workflows/prerelease.yaml
@@ -1,0 +1,56 @@
+name: Create Prerelease Tagged Image
+
+on:
+  release:
+    types: [prereleased]
+
+jobs:
+  rc-release:
+    runs-on: ubuntu-latest
+    env:
+      SHA8: ${GITHUB_SHA::8}
+
+    steps:
+      - name: Checkout opl-api-gateway
+        uses: actions/checkout@v2
+
+      - name: Tag with prereleased version
+        uses: tinact/docker.image-retag@1.0.2
+        with:
+          image_name: openpracticelibrary/opl-api-gateway
+          image_old_tag: $SHA8
+          image_new_tag: ${{ github.event.release.tag_name }}
+          registry: quay.io
+          registry_username: ${{ secrets.QUAY_USER }}
+          registry_password: ${{ secrets.QUAY_PASS }}
+
+      - name: Tag with latest for future full releases
+        uses: tinact/docker.image_retag@1.0.2
+        with:
+          image_name: openpracticelibrary/opl-api-gateway
+          image_old_tag: $SHA8
+          registry: quay.io
+          registry_username: ${{ secrets.QUAY_USER }}
+          registry_password: ${{ secrets.QUAY_PASS }}
+
+  kustomize:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Setup Kustomize
+        uses: imranismail/setup-kustomize@v1
+        with:
+          kustomize-version: "3.5.5"
+
+      - name: Checkout cd repo
+        uses: actions/checkout@v2
+        with:
+          repository: openpracticelibrary/opl-cd
+          token: ${{ secrets.SERVICE_PAT }}
+
+      - name: Kustomize build
+        run: |
+          cd openpracticelibrary/api-gateway/overlays/prerelease
+          kustomize edit set image quay.io/openpracticelibrary/opl-api-gateway:${{ github.event.release.tag_name }}
+          git add .
+          git commit -m "Automated Action: Updated opl-api-gateway prerelease image tag to `${{ github.event.release.tag_name }}`"
+          git push origin master
