+++ Dockerfile
@@ -1,0 +1,8 @@
+FROM node:lts
+WORKDIR /app
+COPY package*.json ./
+RUN npm install
+COPY . .
+EXPOSE 5000
+ENTRYPOINT ["node", "gateway.js"]
+
