+++ gateway.js
@@ -1,7 +1,7 @@
-const { ApolloServer } = require("apollo-server");
-const { ApolloGateway } = require("@apollo/gateway");
+const { ApolloServer } = require('apollo-server');
+const { ApolloGateway } = require('@apollo/gateway');
 
-const { federatedApis } = require("./config");
+const { federatedApis } = require('./src');
 
 const createGateway = (serviceList) => {
   return new ApolloGateway({
@@ -14,19 +14,19 @@ const createGateway = (serviceList) => {
   });
 };
 
-(async () => {
-  const server = new ApolloServer({
-    gateway: createGateway(federatedApis),
-    // If using Apollo Graph Manager, ensure `ENGINE_API_KEY` is set in environment
-    engine: false,
-    // Not currently supported, needs to be explicitly turned off
-    subscriptions: false,
-  });
+const server = new ApolloServer({
+  gateway: createGateway(federatedApis),
+  // If using Apollo Graph Manager, ensure `ENGINE_API_KEY` is set in environment
+  engine: false,
+  // Not currently supported, needs to be explicitly turned off
+  subscriptions: false,
+});
 
-  server.listen().then(({ url }) => {
-    console.log(`🚀 Server ready at ${url}`);
-  });
-})();
+server.listen().then(({ url }) => {
+  console.info(`🚀 Gateway Server ready at ${url}`);
+  console.info('Federated APIs available at:');
+  console.info(federatedApis);
+});
 
 // For testing
 module.exports = createGateway;
