+++ src/index.js
@@ -1,0 +1,3 @@
+const config = require('./config/config');
+
+module.exports = { config };
