+++ test/utils/setEnvVars.js
@@ -1,0 +1,3 @@
+process.env.CONTENT_API_ENDPOINT = "https://not.a.real.url";
+process.env.VERSIONING_API_ENDPOINT = "https://also.not.real";
+process.env.MEDIA_API_ENDPOINT = "https://the.fakest.url";
