+++ test/utils/testServer.js
@@ -1,11 +1,11 @@
-const { ApolloServer } = require('apollo-server');
-const { buildFederatedSchema } = require('@apollo/federation');
+const { ApolloServer } = require("apollo-server");
+const { buildFederatedSchema } = require("@apollo/federation");
 
 const constructTestServer = (typeDefs, resolvers) => {
   const server = new ApolloServer({
     schema: buildFederatedSchema({
       typeDefs,
-      resolvers
+      resolvers,
     }),
     introspection: false,
     playground: false,
